refactor(support): split MarketServiceSupport into single-responsibility classes
- Deleted 663-line god class MarketServiceSupport
- Introduced MarketDtoMapper for all entity-to-DTO conversions
- Introduced MarketPriceCalculator for change/percent calculations
- Introduced KlineDataNormalizer for K-line data parsing
- Introduced MockSectorNetworkGenerator for mock network data
- Moved getHotStocks query logic back into MarketServiceImpl
- Updated MarketFallbackSupport to depend only on the two helpers it needs
- Aligned with ARCHITECTURE-EVALUATION.md recommendations

Closes #448
